forensic_econ_app/requirements.txt
Flask==2.2.3
SQLAlchemy==1.4.46
Flask-SQLAlchemy==2.5.1
python-dateutil==2.8.2
pandas==1.5.3
xlsxwriter==3.0.8
decimal==1.0.0
pytest==7.3.1
